requirements.txt
yt-dlp
faster-whisper
ffmpeg-python